Renamed Asn1SccServiceProviderStub for ParsingServiceProviderStub
